backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.2.0
torchvision==0.17.0
numpy==1.24.3
pandas==2.1.0
Pillow==10.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
scikit-learn==1.3.0
opencv-python==4.8.0.74